Moved NotificationUtils into utils package
